Z probe target ADC value is now a 32 bit integer

+++ src/Endstops/LocalZProbe.cpp
@@ -147,7 +147,7 @@ GCodeResult LocalZProbe::AppendPinNames(const StringRef& str) noexcept
 // Functions used only with scanning Z probes
 float LocalZProbe::GetCalibratedReading() const noexcept
 {
-	const float diff = (float)((int16_t)GetRawReading() - targetAdcValue);
+	const float diff = (float)((int32_t)GetRawReading() - targetAdcValue);
 	return diff * (linearCoefficient + (diff * quadraticCoefficient)) + GetActualTriggerHeight();
 }
